Move the root command to a separate file
This is a prerequisite of #100 because for the `--no-color` flag, we would love to chain up root command's `PersistentPreRun` with all the sub commands. In other words, we would like to execute root command's `PersistentPreRun` before executing any sub command's `PersistentPreRun.

Closes #103.
